removing uneeded state folders and renaming taskscheduler to queue
