+++ src/app/app.component.ts
@@ -3,11 +3,11 @@ import { Subscription } from 'rxjs';
 import { GuessListComponent } from './game/components/guess-list/guess-list.component';
 import { GuessComponent } from './game/components/guess/guess.component';
 import { ProgressBarComponent } from './game/components/progress-bar/progress-bar.component';
-import { WoozleTask } from './task-scheduler/models/woozle-task';
-import { WoozleTaskState } from './task-scheduler/models/woozle-task-state';
+import { WoozleTask } from './queue/models/woozle-task';
+import { WoozleTaskState } from './queue/models/woozle-task-state';
 import { GuessService } from './game/services/guess/guess.service';
 import { PlayerService } from './game/services/player/player.service';
-import { TaskSchedulerService } from './task-scheduler/services/task-scheduler.service';
+import { QueueService } from './queue/services/queue.service';
 
 @Component({
   selector: 'app-root',
@@ -25,11 +25,11 @@ export class AppComponent implements OnDestroy {
   private subscriptions: Subscription[] = [];
 
   constructor(private playerService: PlayerService
-    ,private taskSchedulerService: TaskSchedulerService
+    ,private queueService: QueueService
     ,private guessService: GuessService) {
 
-    this.subscriptions.push(this.taskSchedulerService.taskScheduler$.subscribe((task: WoozleTask) => {
-      if (task.taskState === WoozleTaskState.ENDED && !this.taskSchedulerService.hasRunningTask()) {
+    this.subscriptions.push(this.queueService.taskScheduler$.subscribe((task: WoozleTask) => {
+      if (task.taskState === WoozleTaskState.ENDED && !this.queueService.hasRunningTask()) {
         this.isPlaying = false
       }
     }));
