+++ src/app/game/state/game-state.model.ts
