+++ src/app/queue/models/woozle-task-state.ts
