+++ src/app/queue/state/queue-state.model.ts
@@ -1,0 +1,3 @@
+export interface TaskSchedulerState {
+  
+}
